requirements.txt
openai #==0.27.8
python-dotenv #=1.0.0
fastapi #==0.100.1
uvicorn #==0.23.1
starlette